Nev version 0.2a
The first version of report in the form of a new Calc document.
Partial support for settings. For now the only changeable setting is number of folds in K-fold CV. The setting can be changed directly in the Settings sheet only, the Change settings button just pops-up a message saying so. The style of the settings area in SMLR sheet have been reworked to be operated by the function hooked to the Change settings button.